Try to checkout same branch in dependencies
This implements some behavior from our Jenkins CI that
attemps to checkout a branch of the same name in all
dependency packages. This requires setting `fetch-depth: 0`
in the actions/checkout steps and installing git 2.18+
before those checkout steps. For 18.04, the git-core ppa
is used; this will not be needed on 20.04.